Refactor out legacy head spec
